quantum_annealing/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dimod==0.9.1
dwave-cloud-client==0.7.2
dwave-hybrid==0.4.3
dwave-inspector==0.1.5
dwave-neal==0.5.4
dwave-networkx==0.8.6
dwave-ocean-sdk==2.1.1
dwave-qbsolv==0.3.0
dwave-system==0.9.4
dwave-tabu==0.2.2
dwavebinarycsp==0.1.2
Flask==1.1.2
homebase==1.0.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
minorminer==0.1.9
networkx==2.4
numpy==1.18.3
ortools==7.6.7691
pandas==1.0.3
penaltymodel==0.16.3
penaltymodel-cache==0.4.1
penaltymodel-lp==0.1.3
penaltymodel-mip==0.2.3
plotly==4.6.0
plucky==0.4.3
protobuf==3.11.3
pyqubo==0.4.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
scipy==1.4.1
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
